+++ perfective/disassemble.py
@@ -1,0 +1,4 @@
+import dis
+from utils import validators
+
+dis.dis(validators)